algoristics-be/requirements.txt
pymupdf
uvicorn
openai
python-docx
python-dotenv
pydantic
sentence-transformers
fastapi
starlette
opensearch-py
python-multipart